+++ install/linux/cli/fish.sh
@@ -1,6 +1,17 @@
-sudo apt-add-repository -y ppa:fish-shell/release-3
-sudo apt update
-sudo apt install -y fish
+# Install Fish shell from official PPA
 
-# Set fish as the default shell
-chsh -s /usr/bin/fish
+if ! is_installed fish; then
+  installing_banner "Fish shell"
+  
+  # Add official Fish shell PPA
+  sudo apt-add-repository -y ppa:fish-shell/release-3
+  sudo apt update
+  
+  # Install fish shell
+  apt_install fish
+  
+  # Set fish as the default shell
+  sudo chsh -s /usr/bin/fish "$USER"
+else
+  skipping "Fish shell"
+fi