+++ install/linux/prereq/i3.sh
@@ -6,5 +6,6 @@ source "${DOTFILES_DIR}"/lib/linux.sh
 # Note: the ^^ converts the variable to uppercase
 if [ "${DOTFILES_CONFIG_I3^^}" = "TRUE" ]; then
 	banner "=== Installing i3 Requirements"
-  sudo apt install -y polybar rofi picom redshift xautolock numlockx playerctl yad
+	# Install i3 window manager and related tools
+	apt_install polybar rofi picom redshift xautolock numlockx playerctl yad
 fi